+++ ora_migrator--1.1.0.sql
@@ -51,6 +51,30 @@ $$DECLARE
          '  AND owner NOT IN (' || sys_schemas || E')'
       ')'', max_long ''%s'', readonly ''true'')';
 
+   table_comments text := E'CREATE FOREIGN TABLE %I.table_comments (\n'
+      '   schema     text NOT NULL,\n'
+      '   table_name text NOT NULL,\n'
+      '   table_comment text\n'
+      ') SERVER %I OPTIONS (table ''('
+         'SELECT owner,\n'
+         '       table_name,\n'
+         '       c.comments table_comment\n'
+         'FROM dba_tables t\n'
+         'JOIN dba_tab_comments c\n'
+         '   USING (owner, table_name)\n'
+         'WHERE t.temporary = ''''N''''\n'
+         '  AND t.secondary = ''''N''''\n'
+         '  AND t.nested    = ''''NO''''\n'
+         '  AND t.dropped   = ''''NO''''\n'
+         '  AND (owner, table_name)\n'
+         '     NOT IN (SELECT owner, mview_name\n'
+         '             FROM dba_mviews)\n'
+         '  AND (owner, table_name)\n'
+         '     NOT IN (SELECT log_owner, log_table\n'
+         '             FROM dba_mview_logs)\n'
+         '  AND owner NOT IN (' || sys_schemas || E')'
+      ')'', max_long ''%s'', readonly ''true'')';
+
    columns_sql text := E'CREATE FOREIGN TABLE %I.columns (\n'
       '   schema        text    NOT NULL,\n'
       '   table_name    text    NOT NULL,\n'
@@ -63,8 +87,8 @@ $$DECLARE
       '   nullable      boolean NOT NULL,\n'
       '   default_value text\n'
       ') SERVER %I OPTIONS (table ''('
-         'SELECT col.owner,\n'
-         '       col.table_name,\n'
+         'SELECT owner,\n'
+         '       table_name,\n'
          '       col.column_name,\n'
          '       col.column_id,\n'
          '       CASE WHEN col.data_type_owner IS NULL\n'
@@ -88,14 +112,46 @@ $$DECLARE
          '            FROM dba_views\n'
          '            WHERE owner NOT IN (' || sys_schemas || E')\n'
          '        ) tab\n'
-         '      ON tab.owner = col.owner AND tab.table_name = col.table_name\n'
-         'WHERE (col.owner, col.table_name)\n'
+         '      USING (owner, table_name)\n'
+         'WHERE (owner, table_name)\n'
          '     NOT IN (SELECT owner, mview_name\n'
          '             FROM dba_mviews)\n'
-         '  AND (col.owner, col.table_name)\n'
+         '  AND (owner, table_name)\n'
          '     NOT IN (SELECT log_owner, log_table\n'
-         '             FROM dba_mview_logs)\n'
-         '  AND col.owner NOT IN (' || sys_schemas || E')'
+         '             FROM dba_mview_logs)'
+      ')'', max_long ''%s'', readonly ''true'')';
+
+   column_comments text := E'CREATE FOREIGN TABLE %I.column_comments (\n'
+      '   schema        text    NOT NULL,\n'
+      '   table_name    text    NOT NULL,\n'
+      '   column_name   text    NOT NULL,\n'
+      '   comment       text\n'
+      ') SERVER %I OPTIONS (table ''('
+         'SELECT owner,\n'
+         '       table_name,\n'
+         '       column_name,\n'
+         '       cc.comments\n'
+         'FROM dba_tab_columns col\n'
+         '   JOIN dba_col_comments cc\n'
+         '      USING (owner, table_name, column_name)\n'
+         '   JOIN (SELECT owner, table_name\n'
+         '            FROM dba_tables\n'
+         '            WHERE owner NOT IN (' || sys_schemas || E')\n'
+         '              AND temporary = ''''N''''\n'
+         '              AND secondary = ''''N''''\n'
+         '              AND nested    = ''''NO''''\n'
+         '              AND dropped   = ''''NO''''\n'
+         '         UNION SELECT owner, view_name\n'
+         '            FROM dba_views\n'
+         '            WHERE owner NOT IN (' || sys_schemas || E')\n'
+         '        ) tab\n'
+         '      USING (owner, table_name)\n'
+         'WHERE (owner, table_name)\n'
+         '     NOT IN (SELECT owner, mview_name\n'
+         '             FROM dba_mviews)\n'
+         '  AND (owner, table_name)\n'
+         '     NOT IN (SELECT log_owner, log_table\n'
+         '             FROM dba_mview_logs)'
       ')'', max_long ''%s'', readonly ''true'')';
 
    checks_sql text := E'CREATE FOREIGN TABLE %I.checks (\n'
@@ -727,10 +783,18 @@ BEGIN
    EXECUTE format('DROP FOREIGN TABLE IF EXISTS %I.tables', schema);
    EXECUTE format(tables_sql, schema, server, v_max_long);
    EXECUTE format('COMMENT ON FOREIGN TABLE %I.tables IS ''Oracle tables on foreign server "%I"''', schema, server);
+   /* table comments */
+   EXECUTE format('DROP FOREIGN TABLE IF EXISTS %I.table_comments', schema);
+   EXECUTE format(table_comments, schema, server, v_max_long);
+   EXECUTE format('COMMENT ON FOREIGN TABLE %I.table_comments IS ''Oracle comments for tables on foreign server "%I"''', schema, server);
    /* columns */
    EXECUTE format('DROP FOREIGN TABLE IF EXISTS %I.columns', schema);
    EXECUTE format(columns_sql, schema, server, v_max_long);
    EXECUTE format('COMMENT ON FOREIGN TABLE %I.columns IS ''columns of Oracle tables and views on foreign server "%I"''', schema, server);
+   /* column comments */
+   EXECUTE format('DROP FOREIGN TABLE IF EXISTS %I.column_comments', schema);
+   EXECUTE format(column_comments, schema, server, v_max_long);
+   EXECUTE format('COMMENT ON FOREIGN TABLE %I.column_comments IS ''Oracle comments for columns of tables and views on foreign server "%I"''', schema, server);
    /* checks */
    EXECUTE format('DROP FOREIGN TABLE IF EXISTS %I.checks', schema);
    EXECUTE format(checks_sql, schema, server, v_max_long);