+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   This simplifies the code, and it should be no problem: you can always drop
   and re-create the extension.
 
+- Add foreign tables for table and column comments in the Oracle staging schema.  
+  Patch by Михаил.
+
 # Version 1.0.0, released 2023-02-08 #
 
 ## Enhancements: ##